% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{check_format}
\alias{check_format}
\title{check_format}
\usage{
check_format(raw_data, verbose = FALSE)
}
\arguments{
\item{raw_data}{a data frame containing A&E provider data
for one month, from the NHS England website}

\item{verbose}{control level of detail returned}
}
\value{
boolean indicating whether data frame is in correct format
for analysis. Length 1 if verbose = FALSE, length 6 if not - in this case
each element pertains to a specific column of raw_data that is checked.
}
\description{
check_format
}
