% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{clean_AE_data}
\alias{clean_AE_data}
\title{clean_AE_data}
\usage{
clean_AE_data(raw_data)
}
\arguments{
\item{raw_data}{dataframe containing a NHS England A&E Monthly report
with a standardised set of columns}
}
\value{
the same data as raw_data, as a rectangular table with header removed,
new column names, and correct numerical data types for numerical columns
}
\description{
clean_AE_data
}
