% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{download_AE_files}
\alias{download_AE_files}
\title{download_AE_files}
\usage{
download_AE_files(file_urls, directory)
}
\arguments{
\item{file_urls}{list of urls of files to download}

\item{directory}{location to save files to}
}
\value{
vector of download.file return values
}
\description{
download_AE_files
}
