% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{get_date}
\alias{get_date}
\title{get_date}
\usage{
get_date(raw_data)
}
\arguments{
\item{raw_data}{a data frame containing A&E provider data
for one month, from the NHS England website}
}
\value{
the period (month) that this data pertains to
}
\description{
get_date
}
\examples{
dataList <- load_AE_files(directory = 'nhsAEscraper/sitreps/')
get_date(dataList[[1]])

}
