% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{getAE_data}
\alias{getAE_data}
\title{getAE_data}
\usage{
getAE_data(update_data = TRUE, directory = file.path("data-raw", "sitreps"),
  url_list = NULL)
}
\arguments{
\item{update_data}{whether to download files afresh from NHS England website (TRUE)
or use existing downloaded files (FALSE)}

\item{directory}{directory to find existing downloaded files, and to save new downloads}

\item{url_list}{list of urls (as strings) for the pages to scrape for data files}
}
\value{
A data frame containing all the monthly A&E data from the NHS England website.
}
\description{
getAE_data
}
\examples{
AE_data <- getAE_data(directory = file.path('nhsAEscraper','sitreps'))
str(AE_data)
}
