% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{getAEdata_page_urls_monthly}
\alias{getAEdata_page_urls_monthly}
\title{getAEdata_page_urls_monthly}
\usage{
getAEdata_page_urls_monthly(index_url)
}
\arguments{
\item{index_url}{the url of the page to scrape data files from}
}
\value{
character vector of the urls for NHS England A&E data *.xls files
from one of the monthly data index pages
}
\description{
getAEdata_page_urls_monthly
}
\examples{
getAEdata_page_urls_monthly('https://www.england.nhs.uk/statistics/statistical-work-areas/ae-waiting-times-and-activity/ae-attendances-and-emergency-admissions-2017-18/')
}
