% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{getAEdata_urls_monthly}
\alias{getAEdata_urls_monthly}
\title{getAEdata_urls_monthly}
\usage{
getAEdata_urls_monthly()
}
\value{
the urls for NHS England A&E data *.xls files from two pages
yielding addresses for monthly data from June 2015 to (in principle) present.
}
\description{
getAEdata_urls_monthly
}
\examples{
getAEdata_urls_monthly()
}
