% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{delete_extra_columns}
\alias{delete_extra_columns}
\title{delete_extra_columns}
\usage{
delete_extra_columns(df)
}
\arguments{
\item{df}{a data frame containing A&E provider data
for one month, from the NHS England website}
}
\value{
df with superfluous columns removed
}
\description{
delete_extra_columns
}
