% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AE_data_Eng.R
\name{load_AE_files}
\alias{load_AE_files}
\title{load_AE_files}
\usage{
load_AE_files(directory = "data-raw/sitreps/")
}
\arguments{
\item{directory}{path of the directory to load files from}
}
\value{
a list of data frames containing data loaded from files in directory
whose name is of the form '\*AE-by-provider\*.xls'
}
\description{
load_AE_files
}
